% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_amolz.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot map of accident locations in a year for a state in the US}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric or character string represting a valid state number}

\item{year}{Numeric or Character String represting a calendar year}
}
\value{
Plot a map/ graphic of accident locations for a particular state in the US
}
\description{
This function plots a map representing accident locations for a valid state number
   and a valid year as input
}
\details{
This function, takes a valid state number as input (using \code{state.num}) and
   a valid year (using \code{year}) as inputs,
   creates a filename (using \code{\link{make_filename}}) that represents accidents data in the given year,
   reads the dataframe (using \code{\link{fars_read}}),
   subsets the dataframe for the given state and
   plots the accident locations using Latitude and Longitude data points
}
\note{
Functions stops or shows error message for invalid state number & year or if any argument is left blank
}
\examples{
\dontrun{
fars_map_state(1,2013)
fars_map_state("1","2014")}

}
