% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_amolz.R
\name{fars_read}
\alias{fars_read}
\title{Print glimpse about motor vehicle accidents in the US}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string indicating the file name of dataframe and suffixed with .csv}
}
\value{
If the filename doesn't exist then this function prints that the filename does not exists. Otherwise
   it prints the tibble with first 10 rows and few fields in tabular form and remaining field names in text
}
\description{
\code{fars_read} reads and prints dataframe of the mentioned file name from the
   working directory if it exists
}
\details{
This function checks if the filename (using the \code{filename}) mentioned as the function argument
   exists in the working directory.
}
\note{
filename mentioned without quotes " " will get considered as invalid file arguments
}
\examples{
\dontrun{
fars_read("accident_2013.csv")
fars_read("accident_2015.csv")}

}
