% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_amolz.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read and Print month & year of accident data in the US in given years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years where each year is a 4 digit numeric or character string}
}
\value{
Prints a tibble with Month and year for first 10 rows for each of the valid year inputs
}
\description{
Reads and Prints month and year data in the dataframes from working directory that
   contain fatalities data of accidents in the US in given years
}
\details{
This function takes vector of valid years and applies make_filename
  and fars_read functions to read and print tibble of valid files from working directory
}
\note{
Function returns an error message if the years input is not a valid year
}
\examples{
\dontrun{
fars_read_years(c(2013,2014))
fars_read_years(c(2013,2014, 2015))}

}
