% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_amolz.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize number of observations}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years where each year is a 4 digit numeric or character string}
}
\value{
This function prints a tibble of months as rows, years as fields and number of observations
   as a table
}
\description{
This function summarizes the number of observations representing motor vehicle accidents
   for each month in different years
}
\details{
This function reads one or more accident dataframes using \code{years} as input and summarizes
   in tabular form the number of observations for each month in different years
}
\note{
This function returns an error for invalid years input or no input
}
\examples{
\dontrun{
fars_summarize_years(c(2013))
fars_summarize_years(c(2013, 2014, "2015"))}

}
