% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_amolz.R
\name{make_filename}
\alias{make_filename}
\title{Create and Print file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Numeric or Character String represting a calendar year}
}
\value{
Prints the filename of zipped .csv file
}
\description{
\code{make_filename} creates and prints a filename representing zipped .csv file
   that would contain fatalities data in a given year in the US
}
\details{
This function converts the mentioned argument (using the \code{year}) into an integer
and creates a zipped .csv file name prefixed with "accident_"
}
\note{
leaving argument blank will lead to an missing argument error as there's no default value
}
\examples{
\dontrun{
make_filename(2015)
make_filename("2017")
make_filename(17)}

}
